Token Module introduced (correspond to LRC-2) (#49)
* feat: token module

The major feature in this PR is bellows

- Publish token
  - the token owner publishes the token
  - the token is sent to the owner
- Transfer token
  - An account can transfer token which possessed by the account to another
- Mint token
  - The token owner can mint token
  - The amount is added to the token owner
- Burn token
  - The token owner can burn token
  - The amount is subtracted from the token owner


This PR was supposed to replace the LRC-2.
However, as we discussed, the nature of this project is different from the rationale of the LRC specifications.
I decided to adopt the semantics of the LRC-2 and to ignore the syntax.
It can be LRC-nn. But it is not good time to put this in LRC-nn. Let it just `token` module.

Limitations
- There is no way to define the token owner yet. It means that anybody can publish any token they want.
- When the `bank` module lives, you can transfer token by that module. There is a need to limit the module.