feat: iam module newly introduced (#101)
Introduce Permission Management Module.

The module has features managing this data structure
Address - Permissions{Resource, Action}

The address is the key for permissions.
Permission is an interface which has 2 methods in it.

GetResource() string
GetAction() string
By these 2 methods, permission is defined.
The Module does not support any query and transaction.
The other module which using this module has a responsibility to define the Permission. And the data is stored in the prefixStore so that data from the other modules are separated.